+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 IMPROVEMENTS:
 
+- Improve eip connection ([#21](https://github.com/alibaba/footmark/pull/21))
 - Merge inner_ip and private_ip, and merge eip and public_ip ([#17](https://github.com/alibaba/footmark/pull/17))
 
 BUG FIXES:
 
-- Fix security token not working bug ([#18](https://github.com/alibaba/footmark/pull/18))
+- Fix security token not working bug ([#20](https://github.com/alibaba/footmark/pull/20))
 
 ## 1.8.0 (12 November, 2018)